requirements.txt
tensortrade
pandas_ta
torch
ray[tune]
ray[default]
yfinance
scikit-learn
gymnasium
dm_tree
typer
opencv-python
lz4
stable-baselines3